Refactor code and remove WokwiStartDebugConfig setup
The main action taken in this commit is the refactoring and removal of the `WokwiStartDebugConfig` setup. Deleting these configurations helped fix the issue of the missing running instance of the Wokwi simulator. Furthermore, `WokwiProcessHandler` has been relocated to the `processHandler` package and all its references have been updated.

On a related note, the existing comments were enhanced in `WokwiStartDebugBeforeRunTaskProvider` and `WokwiStartDebugBeforeRunTask` classes to provide better understanding of the debugging process. Lastly, config type references in `WokwiConfigurationFactory` and `plugin.xml` were adjusted to comply with these changes.
