+++ src/main/kotlin/com/github/jozott00/wokwiintellij/execution/processHandler/WokwiProcessHandler.kt
@@ -1,4 +1,4 @@
-package com.github.jozott00.wokwiintellij.execution
+package com.github.jozott00.wokwiintellij.execution.processHandler
 
 import com.github.jozott00.wokwiintellij.simulator.WokwiSimulatorListener
 import com.intellij.execution.process.ProcessHandler
