+++ src/main/kotlin/com/github/jozott00/wokwiintellij/execution/processHandler/WokwiRunProcessHandler.kt
@@ -1,10 +1,8 @@
 package com.github.jozott00.wokwiintellij.execution.processHandler
 
-import com.github.jozott00.wokwiintellij.execution.WokwiProcessHandler
 import com.github.jozott00.wokwiintellij.services.WokwiProjectService
 import com.github.jozott00.wokwiintellij.simulator.EXIT_CODE
 import com.intellij.openapi.components.service
-import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Key
 import java.io.OutputStream
