+++ src/main/resources/META-INF/plugin.xml
@@ -18,7 +18,6 @@
         <notificationGroup id="Wokwi Simulator"
                            displayType="BALLOON"/>
 
-        <configurationType implementation="com.github.jozott00.wokwiintellij.execution.configs.WokwiStartDebugConfigType"/>
         <configurationType implementation="com.github.jozott00.wokwiintellij.execution.configs.WokwiRunConfigType"/>
 
         <stepsBeforeRunProvider
